refactor(flydsl): Merge PA decode host modules
